fix(v1.5.21): add safety gate + zero-CJK detection for DWG CJK recovery (minimal diff from v1.5.19)
Fix 1 - OVER-count protection (巴布亚桩基 59598 vs real 23960):
  NEW shouldReplaceDxfResult() in DwgRawCjkScanner:
    if DXF has >=500 CJK or ratio>=0.15 → PROTECT from byte-scan overwrite
    if recovery > 3.5x DXF count or diversity>=0.7 → REJECT as noise
  Port of desktop wordcount.py:3552 logic

Fix 2 - UNDER-count detection (给排水_t3 Chinese=0 not triggering):
  NEW zeroCjkLoss: total>=500 && itemsCjk<=5 && ratio<0.01
  PDF fallback also triggers on zero-CJK pattern; accepts any PDF with >50 CJK

- Minimal changes on top of v1.5.19 (which compiles) to avoid CI breakage
- versionCode 171 / versionName 1.5.21

+++ app/src/main/java/com/henry/wordcount/DwgRawCjkScanner.kt
@@ -326,7 +326,7 @@ object DwgRawCjkScanner {
         if (dxfCjkCount >= 500 || dxfCjkRatio >= 0.15) return false
         // Recovery 结果膨胀过度 → 拒绝
         if (recovered.cjkTotal > 0 && dxfTotalChars > 0 &&
-            recovered.cjkTotal > dxfTotalChars * MAX_REPLACE_RATIO) return false
+            recovered.cjkTotal.toDouble() > dxfTotalChars * MAX_REPLACE_RATIO) return false
         // Recovery diversity 太高（像随机噪声）→ 拒绝
         if (recovered.cjkDiversity >= 0.7) return false
         // Recovery 有实质内容且看起来像真文本 → 允许
